Rename and move types file to clarify intent
