+++ client/src/static/query/tripQueryTypes.d.ts
@@ -1,4 +1,4 @@
-import { TripQuery } from './gql/graphql.ts';
+import { TripQuery } from '../../gql/graphql.ts';
 
 export type TripPattern = TripQuery['trip']['tripPatterns'][0];
 export type Leg = TripPattern['legs'][0];
